Refactor testAssignSourceDistributionSplits
Move the test into a separate class and remove order dependent
comparison, as the SetMultimap used internally to store assigned splits
doesn't guarantee any particular order.
